Renamed UPnPAuthor to ServiceAuthor, through entire project.
