+++ Tools/ServiceAuthor/ComplexItem.cs
@@ -21,7 +21,7 @@ using System.Collections;
 using System.Windows.Forms;
 using System.ComponentModel;
 
-namespace UPnPAuthor
+namespace ServiceAuthor
 {
 	/// <summary>
 	/// Summary description for ComplexItem.
