+++ Tools/ServiceAuthor/Program.cs
@@ -18,7 +18,7 @@ using System;
 using System.Windows.Forms;
 using System.Collections.Generic;
 
-namespace UPnPAuthor
+namespace ServiceAuthor
 {
     static class Program
     {
